Automatically load the Nerves IEx helpers
This has a few side effects:

1. A `rootfs_overlay` folder is created.
2. `/root/.iex.exs` is no longer read.
3. Some more comments about passing arguments via vm.args were added.

Preventing `/root/.iex.exs` is a safety feature, since errors in that
file caused reboots and could easily brick boards that didn't have
automatic image fallbacks enabled (i.e. all official systems).

The automatical load of the helpers makes the helpers available on the
first boot. It is expected to be safe since `nerves_runtime` is one of
the most exercised Nerves packages.