Generate cookies without symbols
It's frustrating when using Erlang distribution to copy/paste and use
cookies that have symbols in them. They don't copy/paste correctly and
the symbols need to be escaped. To avoid this, change the auto-generated
cookie to be base32 encoded rather than base64.

The resulting cookie has 320 random bits instead of 384. I think that
anyone bothered by this can replace the cookie with their own.