Enable multi_time_warp mode to update Erlang system time
Since it's common for Nerves devices to not have a real-time clock,
Erlang frequently starts up in the 1970s. nerves_time and other
libraries update the OS clock. This time warp is not reflected in the
Erlang system clock by default and that can cause a lot of confusion. To
fix this, enable multi_time_warp mode. See
http://erlang.org/doc/apps/erts/time_correction.html#time-warp-modes.
This mode was chosen over single time warp mode (which would have worked
as well) due to the recommendation from the doc.

Huge thanks to Brian May for figuring out what was going on and why some
libraries would get the incorrect time.