Fix the nerves bootstrap version value
This fixes an issue where the wrong version number was being printed in
error messages and also with `mix nerves.new --version`.

Since the bootstrap version is used in the new project template, it
needs to be modified to just specify the major and minor version.
If not, the generated mix.exs will have a nerves_bootstrap requirement
of "~> x.y.z". That is too strict, so make it "~> x.y".