Disable busy waiting in the BEAM by default
The BEAM will busy wait for a short time before letting the OS scheduler
take over as an optimization to reduce response time latency. The side
effect is to delay other OS processes from running, increase OS-reported
CPU utilization, and possibly increase the processor frequency to deal
with extra load.

Since the advantages of busy waiting are less obvious, this change turns
it off in the generated skeleton project. The intention of the comment
is to direct users to more information so that people who might be
interested in the optimization can turn it back on.