Move core tooling to `:nerves`
`Nerves.Bootstrap` and `Nerves` share a lot of code back and forth
which makes things a little hard to maintain and causes confusion
when trying to following the mix integration with Nerves.

This changes things to move all the tasks to `:nerves` and only use
Nerves.Bootstrap to provided the Nerves project generator and do
the initial loading of `:nerves` into the current Nerves project
being built.

From there, all of the work and heavy lifting with be done with
`:nerves` and maintained there