Move core tooling to `:nerves`
`Nerves.Bootstrap` and `Nerves` share a lot of code back and forth
which makes things a little hard to maintain and causes confusion
when trying to following the mix integration with Nerves.

This changes things to move all the tasks to `:nerves` and only use
Nerves.Bootstrap to provided the Nerves project generator and do
the initial loading of `:nerves` into the current Nerves project
being built.

From there, all of the work and heavy lifting with be done with
`:nerves` and maintained there

The origianl tasks are left in and moved to the `attic` dir for backwards
compatibility. If this bootstrap is used with the new Nerves, module
redefinition warnings are suppresed during compilation so that new
Nerves can redefine these tasks as needed and it will be transparent
to the user.