Precompile nerves packages when loading Nerves
In the old pattern, we would decide that the precompiled pieces of the
tooling were avaialble by checking that `Nerves.Env` module was loaded.
However, with the new tooling, we preemptively load `:nerves` as the
first step so that we can use the tooling there instead of from
Nerves.Bootstrap. This loads `Nerves.Env` and makes these later checks
fail. his means that nerves.loadpaths skips the precompile step, which is
what loads systems, toolchains, and platform into the tooling. This is
crucial because nerves.precompile either resolves an avaialble system
and toolchain to use, or creates an alias as a placeholder to compile
the system to. Without that alias, system compilation fails.

In the new tooling, we skip checking for `Nerves.Env` is loaded and just
run the precompile when we need it. This changes the attic tasks to do
the same thing when running `nerves.loadpaths` so that the right pieces
get added into the tooling stream.