Remove :build_embedded option since it's no longer needed
This option was an early workaround for Makefile projects compiling to a
source priv directory. It was found out that this would always have
problems and that the only fix was to have Makefile projects compile to
the `_build` directory.

The `:build_embedded` option will be removed completely in Elixir 1.15.