Change the erlinit example to be more useful
The current example would set the hostname in the same as the default
for the official Nerves systems. This could be confusing when changing
the hostname via the much more flexible boardid.conf file.

To keep an erlinit example, change it to use `--update-time` to move
forward to erlinits release time. On a device without a real-time clock,
this is a 50 year improvement in the time compared to the default.