Remove legacy tasks
All the mix tooling used to be intermixed between `nerves_bootstrap` and `nerves`
which led to complicated behaviors and confusing upkeep. A while back, the
tooling was moved to `nerves`, but a few mix tasks were kept around here for
backwards compatability.

However, this sometimes gets in the way with Mix, so this removes those tasks
in favor of the `nerves` defined ones