Move Nerves version check to deps.precompile
If a user does not have a `deps` folder with Nerves fetched, then
the new Nerves version check would error out as Nerves.Bootstrap
is starting. However, the compilation would continue and other libs
later on would fail to compile because they would be missing the Nerves
tooling as a result of the bootstraping crash.

To fix this, the check has been moved to the `nerves.bootstrap` task
and also added before `deps.precompile` to ensure the deps are fetched
first before attempting to read any of the `:nerves` dep information