Require the Elixir version that runs mix nerves.new
Since the generated project has a dependency on the Elixir version due
to the vm.args.eex file, set the required Elixir version in the
generated mix.exs. This should hopefully alert a user that something is
amiss if they generate a Nerves project with Elixir 1.17 and go back to
1.16. That way the runtime error due to how IEx is started in the
vm.args.eex will hopefully be less surprising. I.e., we warned you with
the Elixir version warning that something might break. Unfortunately,
this doesn't go the other way with upgrading Elixir versions.