Simplify version matching code
This slightly modifies the semantics for the sake of replacing a
slightly non-obvious function with a call to Version.match?/3. It also
fixes a credo warning about Enum.filter that would have added a couple
lines of code to fix.

The end result is that if you're on a prerelease, you'll get notified to
upgrade to a later prerelease even if it's not the same major, minor,
patch version. This seems like a highly unlikely use case given our
history of releases with this project.